+++ docs/ChangeLog.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow **Source** And **Sink** to discard the stored content.
 - Fresh implementation of application manager.
 - Major API changes in **Reactor** and **Hub**.
+- Major configuration option changes.
 
 ## [15.0.0] - 2025-10-21